+++ packages/base/src/index.ts
@@ -1,2 +1,3 @@
 export * from "./eventStreamSchema";
 export * from "./serialization";
+