+++ packages/fungible-token/src/streams.ts
@@ -1,12 +1,12 @@
 import * as events from "./events";
-import { EventStreamSchema, serializers } from "@proxima-one/stream-schema-base";
-import { BlockchainReference } from "./model";
+import { EventStreamSchema, serializers } from "@proximaone/stream-schema-base";
+import { ChainReference } from "./model";
 
 export type FungibleTokenStreamEvent =
   | events.NewToken
   | events.Approval
   | events.Transfer & {
-  ref?: BlockchainReference
+  ref?: ChainReference
 };
 
 export const fungibleToken: EventStreamSchema<FungibleTokenStreamEvent> = {
@@ -15,13 +15,13 @@ export const fungibleToken: EventStreamSchema<FungibleTokenStreamEvent> = {
   version: "0.1.0",
 }
 
-export type FungibleTokenDiscoveryStreamEvent =
+export type NewFungibleTokenStreamEvent =
   | events.NewToken & {
-  ref?: BlockchainReference
+  ref?: ChainReference
 };
 
-export const fungibleTokenDiscovery: EventStreamSchema<FungibleTokenDiscoveryStreamEvent> = {
-  name: "fungible-token-discovery.streams.proxima.one",
-  serdes: serializers.json<FungibleTokenDiscoveryStreamEvent>(),
+export const newFungibleToken: EventStreamSchema<NewFungibleTokenStreamEvent> = {
+  name: "new-tokens.fungible-token.streams.proxima.one",
+  serdes: serializers.json<NewFungibleTokenStreamEvent>(),
   version: "0.1.0",
 }