+++ packages/base/CHANGELOG.md
@@ -3,7 +3,31 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-## [0.1.2](https://github.com/proxima-one/proxima-npm/compare/@proxima-one/stream-schema-base@0.1.1...@proxima-one/stream-schema-base@0.1.2) (2022-04-28)
+## [0.1.5](https://github.com/proxima-one/proxima-npm/compare/@proximaone/stream-schema-base@0.1.4...@proximaone/stream-schema-base@0.1.5) (2022-03-31)
+
+**Note:** Version bump only for package @proximaone/stream-schema-base
+
+
+
+
+
+## [0.1.4](https://github.com/proxima-one/proxima-npm/compare/@proximaone/stream-schema-base@0.1.3...@proximaone/stream-schema-base@0.1.4) (2022-03-31)
+
+**Note:** Version bump only for package @proximaone/stream-schema-base
+
+
+
+
+
+## 0.1.3 (2022-03-31)
+
+**Note:** Version bump only for package @proximaone/stream-schema-base
+
+
+
+
+
+## [0.1.2](https://github.com/proxima-one/proxima-npm/compare/@proxima-one/stream-schema-base@0.1.1...@proxima-one/stream-schema-base@0.1.2) (2022-02-18)
 
 **Note:** Version bump only for package @proxima-one/stream-schema-base