+++ packages/fungible-token/CHANGELOG.md
@@ -3,7 +3,31 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-## [0.1.5](https://github.com/proxima-one/proxima-npm/compare/@proxima-one/stream-schema-fungible-token@0.1.4...@proxima-one/stream-schema-fungible-token@0.1.5) (2022-04-28)
+## [0.1.8](https://github.com/proxima-one/proxima-npm/compare/@proximaone/stream-schema-fungible-token@0.1.7...@proximaone/stream-schema-fungible-token@0.1.8) (2022-03-31)
+
+**Note:** Version bump only for package @proximaone/stream-schema-fungible-token
+
+
+
+
+
+## [0.1.7](https://github.com/proxima-one/proxima-npm/compare/@proximaone/stream-schema-fungible-token@0.1.6...@proximaone/stream-schema-fungible-token@0.1.7) (2022-03-31)
+
+**Note:** Version bump only for package @proximaone/stream-schema-fungible-token
+
+
+
+
+
+## 0.1.6 (2022-03-31)
+
+**Note:** Version bump only for package @proximaone/stream-schema-fungible-token
+
+
+
+
+
+## [0.1.5](https://github.com/proxima-one/proxima-npm/compare/@proxima-one/stream-schema-fungible-token@0.1.4...@proxima-one/stream-schema-fungible-token@0.1.5) (2022-02-18)
 
 **Note:** Version bump only for package @proxima-one/stream-schema-fungible-token